[dynamo] code style

+++ vertigo-dynamo-impl/src/main/java/io/vertigo/dynamox/domain/constraint/ConstraintIntegerLength.java
@@ -68,7 +68,7 @@ public final class ConstraintIntegerLength extends AbstractConstraintLength<Inte
 		if (value == null) {
 			return true;
 		}
-		final int i = value.intValue();
+		final int i = value;
 		return i > minValue && i < maxValue;
 	}
