+++ vertigo-dynamo-impl/src/main/java/io/vertigo/dynamox/domain/constraint/ConstraintLongLength.java
@@ -66,7 +66,7 @@ public final class ConstraintLongLength extends AbstractConstraintLength<Long> {
 		if (value == null) {
 			return true;
 		}
-		final long i = value.longValue();
+		final long i = value;
 		return i > minValue && i < maxValue;
 	}
